Remove redundant stuff
