+++ packages/hardhat-zksync-solc/src/compile/binary.ts
